requirements.txt
alabaster==0.7.13
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
commonmark==0.9.1
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.4
cycler==0.11.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Django==3.2.19
docutils==0.18.1
fastjsonschema==2.17.1
Flask==2.1.3
Flask-Compress==1.13
fonttools==4.40.0
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
gkeepapi==0.14.2
gpsoauth==1.0.3
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
incremental==22.10.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
keyring==24.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
mistune==2.0.5
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbformat==5.9.0
neo4j-driver==4.4.10
neobolt==1.7.17
packaging==23.1
pandocfilters==1.5.0
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.15.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==1.10.10
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
python-dotenv==1.0.0
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.8
scipy==1.11.1
SecretStorage==3.3.3
service-identity==23.1.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
tenacity==8.2.2
threadpoolctl==3.1.0
tinycss2==1.2.1
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.18
watchdog==3.0.0
Werkzeug==2.0.3
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
